packages/drake-recording-server/pyproject.toml
[project]
name = "drake-recording-server"
version = "0.1.0"
description = "Flask server for recording Drake simulation states as Blender keyframes"
authors = [{ name = "Nicholas Pfaff", email = "nepfaff@mit.edu" }]
readme = "README.md"
license = { text = "BSD-2-Clause" }
requires-python = ">=3.11"
dependencies = [
    "drake>=1.44.0",
    "flask>=3.1.0",
    "numpy>=1.26,<2.0",
    "pillow>=10.0.0",
]

[project.scripts]
drake-recording-server = "drake_recording_server.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/drake_recording_server"]
